be/services/websocketGateway/requirements.txt
flask
flask-socketio
redis
gevent
gunicorn
python-dotenv
pyjwt
flask_cors